Fixed routing issue and moved finance navmenu to finance folder
