+++ src/containers/finances.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import Bills from './bills.js'
 import Payments from './payments.js'
-import FinanceNavmenu from '../components/finance_navbar/finance_navmenu.js'
+import FinanceNavmenu from '../components/finances/finance_navmenu.js'
 
 
 class Finances extends Component {
